5638 update get logs (#83)
update get logs to be:

1) more consistent - use the same syntax consistently
with regards to arrays quotes {} etc.

2) more readable - move more functionality especially
some of the more unique pieces into functions so they
can be more easily used in the future. And separate
pieces of the code by it's goal

3) more extensible - rework some of the more hardcoded
pieces to be more extensible so they can be easily
made to work with future changes

4) more openshift - add support for pieces that were
broken / missing in openshift e.g. routes and dns / etcd

Contributes to: mhub/qp-planning#5638

Signed-off-by: Chris Patmore <christopher.patmore@ibm.com>