5616 zookeeper checks (#88)
sorting doesn't work and is a pain. Basically it doesn't
work for --all-namespaces anyway (which was originally
desired) and it doesn't always work on a single
namespace because some events have null lastTimestamps
which causes it to blow up

begin implementation of ZK checks. To start with just checking each thinks it is okay and finding the leader

add in CR and CRD checks
add some extra resources as per strimzi
getting image tags was added
add init container logs

Contributes to: mhub/qp-planning#5616
Contributes to: mhub/qp-planning#5638

Signed-off-by: Chris Patmore christopher.patmore@ibm.com