feat: Add Release name check to get logs
This commit adds a check of the release name
but performing a list of pods using the release
name as a label selector. If we don't get any
pods returned, then we fail the script as the
release name is wrong or not valid for the
namespace.

It also adds a check of all secrets in the
namespace as not all secrets ES uses have
Release labels associated with them.

We also get the tls.cluster cert which is the
Common Services CA Cert.

Contributes to: mhub/qp-planning#200
Signed-off-by: Tim Mitchell <tim_mitchell@uk.ibm.com>